requirements.txt
gymnasium
stable-baselines3
torch
ultralytics
mss
Pillow
numpy<2
opencv-python
pytesseract
pyautogui